Fixed more type conversion pains

+++ src/diagnostics_component/utils/file_info_utils/inc/file_info_utils.h
@@ -23,7 +23,7 @@ EXTERN_C_BEGIN
  */
 typedef struct tagFileInfo
 {
-    unsigned long fileSize; //!< size of the file in bytes
+    long long fileSize; //!< size of the file in bytes
     char* fileName; //!< the name of the file
     time_t lastWrite; //!< the last time the file was modified
 } FileInfo;
